requirements.txt
asgiref==3.5.2
Django==4.1.3
lxml==4.9.2
Pillow==9.4.0
python-docx==0.8.11
sqlparse==0.4.3
